Cache array indices for molecules
We can use the DynamicArrayObject class to store the location when we
add the molecule to the array.

This requires making setArrayIndex to be made public.  Since it's now
going to be used in multiple places, having it as a protected member
for a friend class doesn't really work any more.